# perf(ext/headers): optimize headers iterable