--- a/ext/fetch/20_headers.js
+++ b/ext/fetch/20_headers.js
@@ -26,7 +26,6 @@ const {
   ArrayPrototypeSort,
   ArrayPrototypeJoin,
   ArrayPrototypeSplice,
-  ObjectEntries,
   ObjectHasOwn,
   RegExpPrototypeTest,
   Symbol,